add gLightWeightScan to help trim down start up - esp. useful for ajax, etc.

+++ BitUser.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * $Header: /cvsroot/bitweaver/_bit_users/BitUser.php,v 1.92 2006/07/28 19:34:18 bitweaver Exp $
+ * $Header: /cvsroot/bitweaver/_bit_users/BitUser.php,v 1.93 2006/07/29 13:56:40 spiderr Exp $
  *
  * Lib for user administration, groups and permissions
  * This lib uses pear so the constructor requieres
@@ -12,7 +12,7 @@
  * All Rights Reserved. See copyright.txt for details and a complete list of authors.
  * Licensed under the GNU LESSER GENERAL PUBLIC LICENSE. See license.txt for details
  *
- * $Id: BitUser.php,v 1.92 2006/07/28 19:34:18 bitweaver Exp $
+ * $Id: BitUser.php,v 1.93 2006/07/29 13:56:40 spiderr Exp $
  * @package users
  */
 
@@ -40,7 +40,7 @@ define("ACCOUNT_DISABLED", -6);
  * Class that holds all information for a given user
  *
  * @author   spider <spider@steelsun.com>
- * @version  $Revision: 1.92 $
+ * @version  $Revision: 1.93 $
  * @package  users
  * @subpackage  BitUser
  */
@@ -178,44 +178,46 @@ class BitUser extends LibertyAttachable {
 
 
 	function updateSession( $pSessionId ) {
+		global $gLightWeightScan;
 		if ( !$this->isDatabaseValid() ) return true;
 		global $gBitSystem, $gBitUser;
 		$update['last_get'] = $gBitSystem->getUTCTime();
 		$update['current_view'] = $_SERVER['PHP_SELF'];
 
-		$this->mDb->StartTrans();
-		$row = $this->mDb->getRow( "SELECT `last_get`, `connect_time`, `get_count`, `user_agent`, `current_view` FROM `".BIT_DB_PREFIX."users_cnxn` WHERE `cookie`=? ", array( $pSessionId ) );
-		if( $gBitUser->isRegistered() ) {
-			$update['user_id'] = $gBitUser->mUserId;
-		}
-		if( $row ) {
-			if( empty( $row['ip'] ) || $row['ip'] != $_SERVER['REMOTE_ADDR'] ) {
-				$update['ip'] = $_SERVER['REMOTE_ADDR'];
+		if( empty( $gLightWeightScan ) ) {
+			$this->mDb->StartTrans();
+			$row = $this->mDb->getRow( "SELECT `last_get`, `connect_time`, `get_count`, `user_agent`, `current_view` FROM `".BIT_DB_PREFIX."users_cnxn` WHERE `cookie`=? ", array( $pSessionId ) );
+			if( $gBitUser->isRegistered() ) {
+				$update['user_id'] = $gBitUser->mUserId;
 			}
-			if( !empty( $_SERVER['HTTP_USER_AGENT'] ) && (empty( $row['user_agent'] ) || $row['user_agent'] != $_SERVER['HTTP_USER_AGENT']) ) {
-				$update['user_agent'] = substr( $_SERVER['HTTP_USER_AGENT'], 0, 128 );
+			if( $row ) {
+				if( empty( $row['ip'] ) || $row['ip'] != $_SERVER['REMOTE_ADDR'] ) {
+					$update['ip'] = $_SERVER['REMOTE_ADDR'];
+				}
+				if( !empty( $_SERVER['HTTP_USER_AGENT'] ) && (empty( $row['user_agent'] ) || $row['user_agent'] != $_SERVER['HTTP_USER_AGENT']) ) {
+					$update['user_agent'] = substr( $_SERVER['HTTP_USER_AGENT'], 0, 128 );
+				}
+				$update['get_count'] = $row['get_count'] + 1;
+				$ret = $this->mDb->associateUpdate( BIT_DB_PREFIX.'users_cnxn', $update, array( 'cookie' => $pSessionId ) );
+			} else {
+				if( $this->isRegistered() ) {
+					$update['user_id'] = $this->mUserId;
+					$update['ip'] = $_SERVER['REMOTE_ADDR'];
+					$update['user_agent'] = substr( $_SERVER['HTTP_USER_AGENT'], 0, 128 );
+					$update['get_count'] = 1;
+					$update['cookie'] = $pSessionId;
+					$result = $this->mDb->associateInsert( BIT_DB_PREFIX.'users_cnxn', $update );
+				}
 			}
-			$update['get_count'] = $row['get_count'] + 1;
-			$ret = $this->mDb->associateUpdate( BIT_DB_PREFIX.'users_cnxn', $update, array( 'cookie' => $pSessionId ) );
-		} else {
-			if( $this->isRegistered() ) {
-				$update['user_id'] = $this->mUserId;
-				$update['ip'] = $_SERVER['REMOTE_ADDR'];
-				$update['user_agent'] = substr( $_SERVER['HTTP_USER_AGENT'], 0, 128 );
-				$update['get_count'] = 1;
-				$update['cookie'] = $pSessionId;
-				$result = $this->mDb->associateInsert( BIT_DB_PREFIX.'users_cnxn', $update );
+			// Delete old connections nightly during the hour of 3 am
+			if( date( 'H' ) == '03' ) {
+				// Default to 30 days history
+				$oldy = $update['last_get'] - ($gBitSystem->getConfig( 'users_cnxn_history_days', 30 ) * 24 * 60);
+				$query = "DELETE from `".BIT_DB_PREFIX."users_cnxn` where `connect_time`<?";
+				$result = $this->mDb->query($query, array($oldy));
 			}
+			$this->mDb->CompleteTrans();
 		}
-		// Delete old connections nightly during the hour of 3 am
-		if( date( 'H' ) == '03' ) {
-			// Default to 30 days history
-			$oldy = $update['last_get'] - ($gBitSystem->getConfig( 'users_cnxn_history_days', 30 ) * 24 * 60);
-			$query = "DELETE from `".BIT_DB_PREFIX."users_cnxn` where `connect_time`<?";
-			$result = $this->mDb->query($query, array($oldy));
-		}
-		$this->mDb->CompleteTrans();
-//die;
 		return true;
 	}
